requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
resend>=0.5.0
schedule>=1.2.0
